+++ src/rleX_Xsl_short_multibyte_encoder.h
@@ -125,7 +125,7 @@ uint32_t CONCAT3(rle, TYPE_SIZE, CONCAT3(_, CODEC, FUNC_NAME))(IN const uint8_t 
 #endif
 #endif
 
-    while (i < inSizeSimd)
+    while ((int64_t)i < inSizeSimd)
     {
 #if defined(IMPL_SSE2) || defined(IMPL_SSSE3)
       const __m128i current = _mm_loadu_si128((const __m128i *)(pIn + i));
@@ -257,7 +257,7 @@ uint32_t CONCAT3(rle, TYPE_SIZE, CONCAT3(_, CODEC, FUNC_NAME))(IN const uint8_t 
       CONCAT3(_rle, TYPE_SIZE, CONCAT3(_, CODEC, process_symbol))(pIn, pOut, i, &state);
 
 #if defined(IMPL_SSE2) || defined(IMPL_SSSE3) || defined(IMPL_AVX2)
-      while (i < inSizeSimd)
+      while ((int64_t)i < inSizeSimd)
       {
 #if defined(IMPL_SSE2) || defined(IMPL_SSSE3)
         const __m128i current = _mm_loadu_si128((const __m128i *)(pIn + i));
